+++ src/pages/Home/CryptoList/CryptoList.scss
@@ -1,0 +1,51 @@
+.crypto-row {
+  display: flex;
+  align-items: center;
+  padding: 0;
+  width: 100%;
+  max-width: 1000px;
+  margin: 5px;
+  min-height: 40px;
+  border-bottom: 1px solid #d2d2d270;
+  cursor: pointer;
+  transition: 0.4s;
+
+  &:hover {
+    transform: translateY(-5px);
+    background: #cdcdcd;
+    box-shadow: 0px 2px 2px rgb(119, 119, 119);
+  }
+
+  &:active {
+    transform: translateY(0px);
+    background: #cdcdcd;
+    box-shadow: 0px 0px 0px rgb(119, 119, 119);
+  }
+
+  &__item {
+    display: flex;
+    align-items: center;
+    margin-left: 20px;
+    width: 20%;
+
+    &-name {
+      font-weight: 600;
+      font-size: 1.2em;
+      text-transform: uppercase;
+      text-shadow: 2px 2px 2px #0095ff;
+    }
+
+    &-symbol {
+      font-weight: 700;
+      color: #7d7d7db6;
+    }
+
+    &-price {
+      font-size: 1.5em;
+    }
+
+    &-market {
+      color: #0095ff;
+    }
+  }
+}
