+++ src/pages/Home/Home.scss
@@ -20,57 +20,17 @@
     &__currency-table {
         max-width: 1000px;
         width: 100%;
-    }
-}
-
-.crypto-row {
-    display: flex;
-    align-items: center;
-    padding: 0;
-    width: 100%;
-    max-width: 1000px;
-    margin: 5px;
-    min-height: 40px;
-    border-bottom: 1px solid #d2d2d270;
-    cursor: pointer;
-    transition: 0.4s;
-
-    &:hover {
-        transform: translateY(-5px);
-        background: #cdcdcd;
-        box-shadow: 0px 2px 2px rgb(119, 119, 119);
+        margin: 0 10px 30px 10px;
     }
 
-    &:active {
-        transform: translateY(0px);
-        background: #cdcdcd;
-        box-shadow: 0px 0px 0px rgb(119, 119, 119);
-    }
-
-    &__item {
+    &__pagination {
         display: flex;
+        justify-content: center;
         align-items: center;
-        margin-left: 20px;
-        width: 20%;
+        margin: 20px;
+    }
 
-        &-name {
-            font-weight: 600;
-            font-size: 1.2em;
-            text-transform: uppercase;
-            text-shadow: 2px 2px 2px #0095ff;
-        }
-
-        &-symbol {
-            font-weight: 700;
-            color: #7d7d7db6;
-        }
-
-        &-price {
-            font-size: 1.5em;
-        }
-
-        &-market {
-            color: #0095ff;
-        }
+    &__pages {
+        margin: 0 10px;
     }
 }
